requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pillow==10.3.0
numpy==1.26.4
onnxruntime==1.18.0
python-multipart==0.0.9